requirements.txt
django
django-webpack-loader==0.7.0

